+++ far/UnicodeString.hpp
@@ -223,8 +223,8 @@ public:
 	friend const UnicodeString operator+(const UnicodeString &strSrc1, const char *lpszSrc2);
 	friend const UnicodeString operator+(const UnicodeString &strSrc1, const wchar_t *lpwszSrc2);
 
-	wchar_t *GetBuffer (int nSize = -1);
-	void ReleaseBuffer (int nLength = -1);
+	wchar_t *GetBuffer (size_t nSize = (size_t)-1);
+	void ReleaseBuffer (size_t nLength = (size_t)-1);
 
 	size_t SetLength(size_t nLength)
 	{
